tests: Cleanup, move and rename sparse tests
