+++ tests/unit/sparse_image.cpp
@@ -16,7 +16,7 @@
 #include "generated/enum_flag_bits.h"
 #include "../framework/layer_validation_tests.h"
 
-TEST_F(NegativeSparse, BindingImageBufferCreate) {
+TEST_F(NegativeSparseImage, BindingImageBufferCreate) {
     TEST_DESCRIPTION("Create buffer/image with sparse attributes but without the sparse_binding bit set");
 
     ASSERT_NO_FATAL_FAILURE(Init());
@@ -73,7 +73,7 @@ TEST_F(NegativeSparse, BindingImageBufferCreate) {
     }
 }
 
-TEST_F(NegativeSparse, ResidencyImageCreateUnsupportedTypes) {
+TEST_F(NegativeSparseImage, ResidencyImageCreateUnsupportedTypes) {
     TEST_DESCRIPTION("Create images with sparse residency with unsupported types");
 
     // Determine which device feature are available
@@ -121,7 +121,7 @@ TEST_F(NegativeSparse, ResidencyImageCreateUnsupportedTypes) {
     CreateImageTest(*this, &image_create_info, "VUID-VkImageCreateInfo-imageType-00972");
 }
 
-TEST_F(NegativeSparse, ResidencyImageCreateUnsupportedSamples) {
+TEST_F(NegativeSparseImage, ResidencyImageCreateUnsupportedSamples) {
     TEST_DESCRIPTION("Create images with sparse residency with unsupported tiling or sample counts");
 
     // Determine which device feature are available
@@ -176,7 +176,7 @@ TEST_F(NegativeSparse, ResidencyImageCreateUnsupportedSamples) {
     CreateImageTest(*this, &image_create_info, "VUID-VkImageCreateInfo-imageType-00976");
 }
 
-TEST_F(NegativeSparse, ResidencyFlag) {
+TEST_F(NegativeSparseImage, ResidencyFlag) {
     TEST_DESCRIPTION("Try to use VkSparseImageMemoryBindInfo without sparse residency flag");
 
     ASSERT_NO_FATAL_FAILURE(Init());
@@ -222,7 +222,7 @@ TEST_F(NegativeSparse, ResidencyFlag) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeSparse, ImageUsageBits) {
+TEST_F(NegativeSparseImage, ImageUsageBits) {
     TEST_DESCRIPTION("Try to use VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT with sparse image");
 
     ASSERT_NO_FATAL_FAILURE(Init());
@@ -253,7 +253,7 @@ TEST_F(NegativeSparse, ImageUsageBits) {
     CreateImageTest(*this, &image_create_info, "VUID-VkImageCreateInfo-None-01925");
 }
 
-TEST_F(NegativeSparse, MemoryBindOffset) {
+TEST_F(NegativeSparseImage, MemoryBindOffset) {
     TEST_DESCRIPTION("Try to use VkSparseImageMemoryBind with offset not less than memory size");
 
     ASSERT_NO_FATAL_FAILURE(Init());
@@ -346,7 +346,7 @@ TEST_F(NegativeSparse, MemoryBindOffset) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(NegativeSparse, QueueBindSparseMemoryType) {
+TEST_F(NegativeSparseImage, QueueBindSparseMemoryType) {
     TEST_DESCRIPTION("Test QueueBindSparse with lazily allocated memory");
 
     ASSERT_NO_FATAL_FAILURE(Init());
@@ -464,207 +464,18 @@ TEST_F(NegativeSparse, QueueBindSparseMemoryType) {
     }
 }
 
-TEST_F(NegativeSparse, QueueBindSparseMemoryBindSize) {
-    TEST_DESCRIPTION("Test QueueBindSparse with invalid sparse memory bind size");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Test requires unsupported sparseBinding feature";
-    }
-
-    VkBufferCreateInfo b_info =
-        vk_testing::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    const auto buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.size = 0u;  // This will trigger the VUID we are testing
-    buffer_memory_bind.memory = buffer_mem.handle();
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info{};
-    buffer_memory_bind_info.buffer = buffer_sparse.handle();
-    buffer_memory_bind_info.bindCount = 1;
-    buffer_memory_bind_info.pBinds = &buffer_memory_bind;
-
-    VkBindSparseInfo bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSparseMemoryBind-size-01098");
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeSparse, QueueBindSparseMemoryBindResourceOffset) {
-    TEST_DESCRIPTION("Test QueueBindSparse with invalid sparse memory bind resource offset");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Test requires unsupported sparseBinding feature";
-    }
-
-    VkBufferCreateInfo b_info =
-        vk_testing::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    const auto buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.size = buffer_mem_reqs.size;
-    buffer_memory_bind.memory = buffer_mem.handle();
-    // This will trigger the VUID we are testing
-    buffer_memory_bind.resourceOffset = buffer_mem_reqs.size + buffer_mem_reqs.size;
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info{};
-    buffer_memory_bind_info.buffer = buffer_sparse.handle();
-    buffer_memory_bind_info.bindCount = 1;
-    buffer_memory_bind_info.pBinds = &buffer_memory_bind;
-
-    VkBindSparseInfo bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSparseMemoryBind-resourceOffset-01099");
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeSparse, QueueBindSparseMemoryBindSizeResourceOffset) {
-    TEST_DESCRIPTION("Test QueueBindSparse with invalid sparse memory bind size due to resource offset");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Test requires unsupported sparseBinding feature";
-    }
-
-    VkBufferCreateInfo b_info =
-        vk_testing::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    const auto buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.memory = buffer_mem.handle();
-    // This will trigger the VUID we are testing
-    buffer_memory_bind.resourceOffset = 100;
-    buffer_memory_bind.size = buffer_mem_reqs.size;
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info{};
-    buffer_memory_bind_info.buffer = buffer_sparse.handle();
-    buffer_memory_bind_info.bindCount = 1;
-    buffer_memory_bind_info.pBinds = &buffer_memory_bind;
-
-    VkBindSparseInfo bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSparseMemoryBind-size-01100");
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeSparse, QueueBindSparseMemoryBindSizeMemoryOffset) {
-    TEST_DESCRIPTION("Test QueueBindSparse with invalid sparse memory bind size due to memory offset");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Test requires unsupported sparseBinding feature";
-    }
-
-    VkBufferCreateInfo b_info =
-        vk_testing::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    const auto buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.memory = buffer_mem.handle();
-    // This will trigger the VUID we are testing
-    buffer_memory_bind.memoryOffset = buffer_mem_alloc.allocationSize - 1;
-    buffer_memory_bind.size = buffer_mem_reqs.size;
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info{};
-    buffer_memory_bind_info.buffer = buffer_sparse.handle();
-    buffer_memory_bind_info.bindCount = 1;
-    buffer_memory_bind_info.pBinds = &buffer_memory_bind;
-
-    VkBindSparseInfo bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSparseMemoryBind-size-01102");
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(NegativeSparse, ImageMemoryBind) {
+TEST_F(NegativeSparseImage, ImageMemoryBind) {
     TEST_DESCRIPTION("Try to bind sparse resident image with invalid VkSparseImageMemoryBind");
 
     ASSERT_NO_FATAL_FAILURE(Init());
 
     if (!m_device->phy().features().sparseBinding || !m_device->phy().features().sparseResidencyImage3D) {
         GTEST_SKIP() << "sparseBinding && sparseResidencyImage3D features are required.";
+    }
+
+    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
+    if (!sparse_index) {
+        GTEST_SKIP() << "Required queue families not present";
     }
 
     VkImageCreateInfo create_info = vk_testing::Image::create_info();
@@ -712,10 +523,6 @@ TEST_F(NegativeSparse, ImageMemoryBind) {
     bind_info.imageBindCount = 1u;
     bind_info.pImageBinds = &image_bind_info;
 
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
     VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
 
     // Force offset.x to invalid value
@@ -780,342 +587,4 @@ TEST_F(NegativeSparse, ImageMemoryBind) {
     vk::QueueBindSparse(sparse_queue, 1, &bind_info, VK_NULL_HANDLE);
     m_errorMonitor->VerifyFound();
     image_bind.subresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-}
-
-TEST_F(NegativeSparse, OverlappingBufferCopy) {
-    TEST_DESCRIPTION("Test overlapping sparse buffers' copy with overlapping device memory");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Requires unsupported sparseBinding feature.";
-    }
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-
-    auto s_info = LvlInitStruct<VkSemaphoreCreateInfo>();
-    vk_testing::Semaphore semaphore(*m_device, s_info);
-
-    VkBufferCopy copy_info;
-    copy_info.srcOffset = 0;
-    copy_info.dstOffset = 0;
-    copy_info.size = 256;
-
-    VkBufferCreateInfo b_info = vk_testing::Buffer::create_info(
-        copy_info.size, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-    VkBufferObj buffer_sparse2;
-    buffer_sparse2.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    VkMemoryAllocateInfo buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-    vk_testing::DeviceMemory buffer_mem2;
-    buffer_mem2.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.size = buffer_mem_reqs.size;
-    buffer_memory_bind.memory = buffer_mem.handle();
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_infos[2] = {};
-    buffer_memory_bind_infos[0].buffer = buffer_sparse.handle();
-    buffer_memory_bind_infos[0].bindCount = 1;
-    buffer_memory_bind_infos[0].pBinds = &buffer_memory_bind;
-    buffer_memory_bind_infos[1].buffer = buffer_sparse2.handle();
-    buffer_memory_bind_infos[1].bindCount = 1;
-    buffer_memory_bind_infos[1].pBinds = &buffer_memory_bind;
-
-    auto bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 2;
-    bind_info.pBufferBinds = buffer_memory_bind_infos;
-    bind_info.signalSemaphoreCount = 1;
-    bind_info.pSignalSemaphores = &semaphore.handle();
-
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-    VkFenceObj sparse_queue_fence(*m_device);
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, sparse_queue_fence);
-    ASSERT_VK_SUCCESS(sparse_queue_fence.wait(kWaitTimeout));
-    // Set up complete
-
-    m_commandBuffer->begin();
-    // This copy is not legal since both buffers share same device memory range, and none of them will be rebound
-    // to non overlapping device memory ranges. Reported at queue submit
-    vk::CmdCopyBuffer(m_commandBuffer->handle(), buffer_sparse.handle(), buffer_sparse2.handle(), 1, &copy_info);
-    m_commandBuffer->end();
-
-    // Submitting copy command with overlapping device memory regions
-    VkPipelineStageFlags mask = VK_PIPELINE_STAGE_ALL_COMMANDS_BIT;
-    auto submit_info = LvlInitStruct<VkSubmitInfo>();
-    submit_info.waitSemaphoreCount = 1;
-    submit_info.pWaitSemaphores = &semaphore.handle();
-    submit_info.pWaitDstStageMask = &mask;
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_commandBuffer->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdCopyBuffer-pRegions-00117");
-    vk::QueueSubmit(m_device->m_queue, 1, &submit_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-
-    // Wait for operations to finish before destroying anything
-    vk::QueueWaitIdle(m_device->m_queue);
-}
-
-TEST_F(NegativeSparse, OverlappingBufferCopy2) {
-    TEST_DESCRIPTION("Test overlapping sparse buffers' copy with overlapping device memory");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Requires unsupported sparseBinding feature.";
-    }
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-
-    auto s_info = LvlInitStruct<VkSemaphoreCreateInfo>();
-    vk_testing::Semaphore semaphore(*m_device, s_info);
-
-    constexpr VkDeviceSize copy_size = 16;
-    std::array<VkBufferCopy, 4> copy_info_list = {};
-    copy_info_list[0].srcOffset = 0;
-    copy_info_list[0].dstOffset = 16;
-    copy_info_list[0].size = copy_size;
-
-    copy_info_list[1].srcOffset = 16;  // source overlaps copy_info_list[0].dst
-    copy_info_list[1].dstOffset = 32;
-    copy_info_list[1].size = copy_size;
-
-    copy_info_list[2].srcOffset = 32;  // source overlaps copy_info_list[1].dst
-    copy_info_list[2].dstOffset = 48;
-    copy_info_list[2].size = copy_size;
-
-    copy_info_list[3].srcOffset = 48;  // source overlaps copy_info_list[2].dst
-    copy_info_list[3].dstOffset = 64;
-    copy_info_list[3].size = copy_size;
-
-    VkBufferCreateInfo b_info =
-        vk_testing::Buffer::create_info(256, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    b_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, b_info);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    VkMemoryAllocateInfo buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem;
-    buffer_mem.init(*m_device, buffer_mem_alloc);
-
-    VkSparseMemoryBind buffer_memory_bind_1 = {};
-    buffer_memory_bind_1.size = buffer_mem_reqs.size;
-    buffer_memory_bind_1.memory = buffer_mem.handle();
-
-    std::array<VkSparseBufferMemoryBindInfo, 1> buffer_memory_bind_infos = {};
-    buffer_memory_bind_infos[0].buffer = buffer_sparse.handle();
-    buffer_memory_bind_infos[0].bindCount = 1;
-    buffer_memory_bind_infos[0].pBinds = &buffer_memory_bind_1;
-
-    auto bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = size32(buffer_memory_bind_infos);
-    bind_info.pBufferBinds = buffer_memory_bind_infos.data();
-    bind_info.signalSemaphoreCount = 1;
-    bind_info.pSignalSemaphores = &semaphore.handle();
-
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-    VkFenceObj sparse_queue_fence(*m_device);
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, sparse_queue_fence);
-    ASSERT_VK_SUCCESS(sparse_queue_fence.wait(kWaitTimeout));
-    // Set up complete
-
-    m_commandBuffer->begin();
-    vk::CmdCopyBuffer(m_commandBuffer->handle(), buffer_sparse.handle(), buffer_sparse.handle(), size32(copy_info_list),
-                      copy_info_list.data());
-    m_commandBuffer->end();
-
-    // Submitting copy command with overlapping device memory regions
-    VkPipelineStageFlags mask = VK_PIPELINE_STAGE_ALL_COMMANDS_BIT;
-    auto submit_info = LvlInitStruct<VkSubmitInfo>();
-    submit_info.waitSemaphoreCount = 1;
-    submit_info.pWaitSemaphores = &semaphore.handle();
-    submit_info.pWaitDstStageMask = &mask;
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_commandBuffer->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdCopyBuffer-pRegions-00117");
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdCopyBuffer-pRegions-00117");
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdCopyBuffer-pRegions-00117");
-    vk::QueueSubmit(m_device->m_queue, 1, &submit_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-
-    // Wait for operations to finish before destroying anything
-    vk::QueueWaitIdle(m_device->m_queue);
-}
-
-TEST_F(NegativeSparse, OverlappingBufferCopy3) {
-    TEST_DESCRIPTION("Test coyping from a range that spans two different memory chunks");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    if (!m_device->phy().features().sparseBinding) {
-        GTEST_SKIP() << "Requires unsupported sparseBinding feature.";
-    }
-
-    const std::optional<uint32_t> sparse_index = m_device->QueueFamilyMatching(VK_QUEUE_SPARSE_BINDING_BIT, 0u);
-    if (!sparse_index) {
-        GTEST_SKIP() << "Required queue families not present";
-    }
-
-    auto semaphore_ci = LvlInitStruct<VkSemaphoreCreateInfo>();
-    vk_testing::Semaphore semaphore(*m_device, semaphore_ci);
-
-    VkBufferCreateInfo buffer_ci =
-        vk_testing::Buffer::create_info(4096 * 32, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT, nullptr);
-    buffer_ci.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    VkBufferObj buffer_sparse;
-    buffer_sparse.init_no_mem(*m_device, buffer_ci);
-
-    VkMemoryRequirements buffer_mem_reqs;
-    vk::GetBufferMemoryRequirements(device(), buffer_sparse.handle(), &buffer_mem_reqs);
-    if (buffer_mem_reqs.alignment <= 1) {
-        GTEST_SKIP() << "Buffer copy will not work as intended if VkMemoryRequirements::alignment is not superior to 1";
-    }
-    buffer_sparse.destroy();
-    buffer_ci.size = 2 * buffer_mem_reqs.alignment;
-    buffer_sparse.init_no_mem(*m_device, buffer_ci);
-    VkMemoryAllocateInfo buffer_mem_alloc =
-        vk_testing::DeviceMemory::get_resource_alloc_info(*m_device, buffer_mem_reqs, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
-
-    vk_testing::DeviceMemory buffer_mem_1;
-    buffer_mem_1.init(*m_device, buffer_mem_alloc);
-    vk_testing::DeviceMemory buffer_mem_2;
-    buffer_mem_2.init(*m_device, buffer_mem_alloc);
-
-    std::array<VkSparseMemoryBind, 2> buffer_memory_binds = {};
-    buffer_memory_binds[0].size = buffer_mem_reqs.alignment;
-    buffer_memory_binds[0].memory = buffer_mem_1.handle();
-    buffer_memory_binds[1].resourceOffset = buffer_mem_reqs.alignment;
-    buffer_memory_binds[1].size = buffer_mem_reqs.alignment;
-    buffer_memory_binds[1].memory = buffer_mem_2.handle();
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info = {};
-    buffer_memory_bind_info.buffer = buffer_sparse.handle();
-    buffer_memory_bind_info.bindCount = size32(buffer_memory_binds);
-    buffer_memory_bind_info.pBinds = buffer_memory_binds.data();
-
-    auto bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-    bind_info.signalSemaphoreCount = 1;
-    bind_info.pSignalSemaphores = &semaphore.handle();
-
-    VkQueue sparse_queue = m_device->graphics_queues()[sparse_index.value()]->handle();
-    VkFenceObj sparse_queue_fence(*m_device);
-    vk::QueueBindSparse(sparse_queue, 1, &bind_info, sparse_queue_fence);
-    ASSERT_VK_SUCCESS(sparse_queue_fence.wait(kWaitTimeout));
-    // Set up complete
-
-    VkBufferCopy copy_info;
-    copy_info.srcOffset = 0;                              // srcOffset is the start of buffer_mem_1, or 0 in this space.
-    copy_info.dstOffset = buffer_mem_reqs.alignment / 2;  // dstOffset is the start of buffer_mem_2, or 0 in this space
-                                                          // => since overlaps are computed in buffer space, none should be detected
-    copy_info.size = buffer_mem_reqs.alignment;
-    m_commandBuffer->begin();
-    vk::CmdCopyBuffer(m_commandBuffer->handle(), buffer_sparse.handle(), buffer_sparse.handle(), 1, &copy_info);
-    m_commandBuffer->end();
-
-    // Submitting copy command with overlapping device memory regions
-    VkPipelineStageFlags mask = VK_PIPELINE_STAGE_ALL_COMMANDS_BIT;
-    auto submit_info = LvlInitStruct<VkSubmitInfo>();
-    submit_info.waitSemaphoreCount = 1;
-    submit_info.pWaitSemaphores = &semaphore.handle();
-    submit_info.pWaitDstStageMask = &mask;
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_commandBuffer->handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdCopyBuffer-pRegions-00117");
-    vk::QueueSubmit(m_device->m_queue, 1, &submit_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-
-    // Wait for operations to finish before destroying anything
-    vk::QueueWaitIdle(m_device->m_queue);
-    vk::QueueWaitIdle(sparse_queue);
-}
-
-TEST_F(NegativeSparse, BufferFlagsFeature) {
-    TEST_DESCRIPTION("Create buffers with Flags that require disabled sparse features");
-
-    VkPhysicalDeviceFeatures features = {};
-    features.sparseBinding = VK_FALSE;
-    features.sparseResidencyBuffer = VK_FALSE;
-    features.sparseResidencyAliased = VK_FALSE;
-
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    ASSERT_NO_FATAL_FAILURE(InitState(&features));
-
-    auto buffer_create_info = LvlInitStruct<VkBufferCreateInfo>();
-    buffer_create_info.size = 64;
-    buffer_create_info.usage = VK_BUFFER_USAGE_TRANSFER_SRC_BIT;
-
-    buffer_create_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    CreateBufferTest(*this, &buffer_create_info, "VUID-VkBufferCreateInfo-flags-00915");
-
-    buffer_create_info.flags = VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkBufferCreateInfo-flags-00916");
-    CreateBufferTest(*this, &buffer_create_info, "VUID-VkBufferCreateInfo-flags-00918");
-
-    buffer_create_info.flags = VK_BUFFER_CREATE_SPARSE_ALIASED_BIT;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkBufferCreateInfo-flags-00917");
-    CreateBufferTest(*this, &buffer_create_info, "VUID-VkBufferCreateInfo-flags-00918");
-}
-
-TEST_F(NegativeSparse, VkSparseMemoryBindMemory) {
-    TEST_DESCRIPTION("test VkSparseMemoryBind::memory is valid");
-
-    ASSERT_NO_FATAL_FAILURE(Init());
-
-    VkBufferCreateInfo buffer_create_info = LvlInitStruct<VkBufferCreateInfo>();
-    buffer_create_info.usage = VK_BUFFER_USAGE_TRANSFER_SRC_BIT;
-    buffer_create_info.size = 1024;
-    buffer_create_info.queueFamilyIndexCount = 0;
-
-    if (m_device->phy().features().sparseResidencyBuffer) {
-        buffer_create_info.flags = VK_BUFFER_CREATE_SPARSE_RESIDENCY_BIT | VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    } else {
-        GTEST_SKIP() << "Test requires unsupported sparseResidencyBuffer feature";
-    }
-
-    VkBufferObj buffer;
-    buffer.init_no_mem(*m_device, buffer_create_info);
-
-    VkDeviceMemory bad_memory = CastToHandle<VkDeviceMemory, uintptr_t>(0xbaadbeef);
-
-    VkSparseMemoryBind buffer_memory_bind = {};
-    buffer_memory_bind.size = 256;
-    buffer_memory_bind.memory = bad_memory;
-    buffer_memory_bind.memoryOffset = 0;
-
-    VkSparseBufferMemoryBindInfo buffer_memory_bind_info = {};
-    buffer_memory_bind_info.buffer = buffer.handle();
-    buffer_memory_bind_info.bindCount = 1;
-    buffer_memory_bind_info.pBinds = &buffer_memory_bind;
-
-    VkBindSparseInfo bind_info = LvlInitStruct<VkBindSparseInfo>();
-    bind_info.bufferBindCount = 1;
-    bind_info.pBufferBinds = &buffer_memory_bind_info;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSparseMemoryBind-memory-parameter");
-    vk::QueueBindSparse(m_device->m_queue, 1, &bind_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
 }
